[FE-201] refactor: 마이 레코드 타입 분리와 리팩토링
