+++ src/types/recordData.ts
@@ -1,3 +1,5 @@
+import { IMemoryRecord } from './myRecord'
+
 export interface IRecordDataType {
   recordId: number
   categoryId: number
@@ -11,21 +13,11 @@ export interface IRecordDataType {
   imageUrls: string[]
 }
 
-export interface IMemoryRecordList {
-  memoryRecordList: IMemoryRecord[]
-  totalCount: number
-  totalPage: number
-}
-
-interface RecordCategory {
+export interface RecordCategory {
   recordId: number
   title: string
   iconName: string
   iconColor: string
-}
-
-export interface IMemoryRecord extends RecordCategory {
-  memoryRecordComments: IRecordMemoryComment[]
 }
 
 export interface CategoryCard {
@@ -34,33 +26,6 @@ export interface CategoryCard {
   iconName: string
   recordId: number
   title: string
-}
-
-export interface IRecordMemoryComment {
-  commentId: number
-  content: string
-}
-
-export interface IRecordByDate {
-  recordId: number
-  title: string
-  categoryName: string
-  commentCount: number
-  iconName: string
-  colorName: string
-  createdAt: string
-}
-
-export interface IRecordByDateList {
-  recordByDateDtos: IRecordByDate[]
-  totalCount: number
-  totalPage: number
-}
-
-export interface IMyRecordRequestParam {
-  date: string
-  page: number
-  size: number
 }
 
 export interface IRandomRecordData
